clickhouse/src/requirements.txt
clickhouse-connect
python-dotenv
